api/requirements.txt
# Framework web
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de données et ORM
sqlalchemy==2.0.23
alembic==1.12.1

# Authentification et sécurité
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Modèles et validation
pydantic==2.5.0
pydantic-settings==2.0.3
email-validator==2.1.0

# Machine Learning et traitement d'images
tensorflow==2.12.0
pillow==10.1.0
numpy==1.24.3

# Middleware et outils
python-dotenv==1.0.0
loguru==0.7.2

# APIs externes
openai==1.3.8
anthropic==0.7.7
httpx==0.25.2

# Monitoring et métriques
prometheus-client==0.19.0

# Développement et tests
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8 